Reflection for automatic reassociation
This commit completes the (first version of) a reflective procedure for
automatically reassociating chains of compositions. This is implemented
via a macro which, given two terms like

  t1 := f . g . ((h . i) . j)
  t2 := (f . g) . h . i . j

constructs a proof that t1 = t2.

The method is still quite rough around the edges. In particular:

- We have no good error reporting.
- There are a couple of TERMINATING pragmas, at least some of which are
  unnecessary.
- The macro doesn't work well with the equational reasoning module from
  the standard library. This is because the combinators in that module
  will typically yield terms of the form `t = ?`, where t is a term but
  ? is an unsolved meta. Our macro then doesn't have enough information
  to work with and fails.

As a side product of implementing reflection support, we don't represent
composition chains as lists any more but rather as trees. This is a much
more natural representation.