Cat: Move module parameters into anonymous module
The top-level module Cats.Category.Cat used to carry some module
parameter. I now believe that this is generally a bad idea, so this
commit moves the parameters from the top-level module to an anonymous
one within it. While we're at it, we make the parameters explicit rather
than implicit, which is more user-friendly anyway.