Sets: monos/epis correspond to inj/surj functions
This commit adds proofs of the facts that
(a) the monos in Sets are exactly the injective functions;
(b) every surjective function is epic in Sets.

The second part of the equivalence between epis and surjective functions
(all epis are surjective) only has a nonconstructive proof as far as I
can tell, so we omit it.