Change cod of Hom functor from Sets to Setoids
Our version of the representable (Homset) functor used to have type
C -> Sets (for some category C). This meant that equality of arrows in C
was required to be propositional equality, which is of course not what
we want for many categories. Hence, this commit changes the codomain to
Setoids, which lets us use C's definition of equality between arrows.