Prepare to break up Category module
The Category module is getting a little unwieldy, so I'd like to break
out the various independent concepts (initial/terminal objects,
epis/monos, products, ...) into their own modules.

In order to do this in a way that doesn't break clients of the Category
module, we need to proceed as follows:

1. Move the Category definition to a new module Category.Base.
2. Import Category.Base from Category and effectively re-export it.
3. Move the various definitons into their own modules, which will depend
   only on Category.Base.
4. Import the definition modules in Category and re-export them so that
   clients can still get all the definitions from one place.

This commit performs steps 1 and 2.