Add sugar for conversion to objects/arrows
The library defines a number of notions -- currently equalizers, isos,
products and statements of unique existence -- which may be regarded as
an object or arrow paired with proofs of some properties. We define a
lightweight notation (via a typeclass in Cats.Util.Conv) for extracting
the defining object/arrow from such a thing.