Generalise levels of Cat operations
The composition and identity operations for functors now admit source
and target categories at different levels. Unfortunately, we still need
source and target categories when viewing functors as arrows in Cat (I
think).

The motivation for this is that the Hom functor has source and target
categories at different levels, so the definition of preservation of
limits needs to be level-generalised as well.

One should consider whether it suffices to have one level per category
(rather than one for objects, arrows and equality respecively). The
current formulation leads to a fair amount of boilerplate.