Level-generalise HasProducts->HasBinaryProducts
The proof that a category has binary products if it has general products
previously required that the general products be at level zero. Since we
have many categories where that is not the case, we generalise the
definition to accept products at any level in the premise (at the
expense of some serious incidental complexity).