Make assoc! macro work with equational reasoning
The assoc! macro automatically solves goals of the form f ~~ g, where f
and g are arrows in some category which are equal by associativity of
composition.

Previously, this macro did not work well with the equational reasoning
combinators. This is because the macro would be run when the goal
equality was still an unsolved meta. Now we force all metas in the goal
term to be solved before the tactic runs. (I'm not sure if this loops if
the goal term contains metas that cannot be solved when the tactic is
run.)