Add more convenient representation of natural isos
A natural iso can now be constructed by giving a family of isos between
components and a proof that each of the isos is natural in the forward
direction. From this, it already follows that each of the isos is also
natural in the backward direction, so we have a natural iso.