Replace assoc! macro with simplification (simp!)
We rewrite the assoc! macro and extend it with a simplification
algorithm that normalises the two sides of an equation before comparing
them. For now, the simplification just transforms `id . f` and `f . id`
into `f`, so it's not very useful. I plan to extend the algorithm to
handle fmap as well, at which point it'll replace a bunch of
boilerplate.